Refactoring of the addToCart process: generalization of the function responsible for adding and separating functions into its own file
